pyproject.toml
[project]
name = "outlook-graph-mcp"
version = "1.1.0"
description = "MCP server for Microsoft Outlook via Microsoft Graph API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Michael Palermiti" }
]
keywords = ["mcp", "outlook", "microsoft-graph", "email", "calendar", "openclaw"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Email",
    "Topic :: Office/Business",
]
dependencies = [
    "mcp[cli]>=1.27",
    "msgraph-sdk>=1.55",
    "azure-identity>=1.25",
    "pydantic>=2.0",
]

[project.urls]
Homepage = "https://github.com/mpalermiti/outlook-mcp"
Repository = "https://github.com/mpalermiti/outlook-mcp"
Issues = "https://github.com/mpalermiti/outlook-mcp/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
outlook-mcp = "outlook_mcp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/outlook_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "src/outlook_mcp",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require real Azure AD auth (deselect with '-m \"not integration\"')",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
]
